Temporarily fixed the ionIoni -problem
There was a bug in the original Hadrontherapy code: the name of the phantom
volume was written as "PhantomPhys" but the correct form is "phantomPhys".
We fixed this bug. Alas, this fix triggered another problem in the original
hadrontherapy code... The error messages on the screen are caused by the line:
 G4cout << "Warning! Unknown proton process: "<< process << G4endl;
in HadrontherapySteppingAction.cc

The relevance of this error message should be investigated further...