Bugfix: AnalysisManager->book() call was commented
This fixes error:
Error in <TROOT::WriteTObject>: The current directory (root) is not associated
with a file. The object (theROOTNtuple) has not been written.
Error in <TROOT::WriteTObject>: The current directory (root) is not associated
with a file. The object (theROOTIonTuple) has not been written.
Error in <TROOT::WriteTObject>: The current directory (root) is not associated
with a file. The object (ThinTargetBeam) has not been written.
Warning in <TFile::Write>: file DoseDistribution.root not opened in write mode
Graphics systems deleted.

The bug was introduced in commit 238c44397e7f788051dc51b7400d7a9479709d08